test(opencode): use real MCP recovery server
